Masters Dissertation
The three main categories of research in cryptocurrency trading are market forecasting models, portfolio optimisation and automated trading. The most popular models being Neural Networks and Deep Learning (DL) among trading algorithms and reinforcement learning and fuzzy logic for automated trading, Nguyen and Chan, (2024). Forecasting in traditional and cryptocurrency markets is a difficult task due to the complexity of the data including random noise made by external market participants Lee and Kim, (2020). Financial time series (FTS) forecasting is a key focus within the financial sector, offering insights into potential risks, trends, and entry and exits for investing. Advances in technology and the continued rise in volume of data have led to more complex models for predicting these signals. Small improvements in forecasting can result in substantial gains. Enhancing the accuracy of these predictions and protecting against market volatility is of great benefit to organisations and individual investors involved in the financial market, Tang et al (2022). How complex can these models get and does this add to their predictive power? There is a push and pull over simplicity over complexity in FTS research. Too simple doesn’t capture non-linearity in the data and too complex is computationally inefficient and risks overfitting. If a model is effective at predicting the direction of an asset but can’t be deployed efficiently, then it can’t be used effectively in real world scenarios. The stock market originated to release funds for businesses and expand production. They act as a form of investment for individuals and institutions who believe in the product or service they represent. Stocks and shares are a biproduct of the industrial revolution where money was needed to fund increased production. The investors received a share of the profits and an increase in their capital invested, Shadbolt, J. et al (2002). Thereafter, came government bonds, aggregates of stock prices (S&P 500, FTSE 100) and complex financial instruments such as futures. These are key components of economics, the financial industry, and people's own personal investment strategies, but they come with risk and uncertainty.

This thesis compares a variety of different architectures and their accuracy in predicting next day prices of ethereum.